Added a custom eslint rule no-dupe-func
The `no-dupe-func` rule works much like the `no-redeclare` rule
except that it works when assigning a function to an object property.

For instance, the following will be marked as an error:

```javascript
let proto = {};

proto.init = function () {
    // original init
};

proto.init = function () {
    // accidental override of init
};
```

The second definition for `proto.init` will give a linting error.