Fixed bug in CLI error processing
It turns out I had the name of the event wrong. I was using
`unhandledException` instead of `uncaughtException` and we didn't
have any tests to exercise the proper display of errors.

I've added the test, and made the fix. We should get actual error
messages now, instead of the cryptic

```
path/to/project/node_modules/beaker/src/cli/utils.js:21
            exitCode: exitCode,
                      ^
```